requirements.txt
streamlit
torch
torchvision
Pillow
albumentations
numpy
matplotlib
opencv-python
pycocotools
